+++ spyro/abc/abc_layer.py
@@ -4,7 +4,7 @@ from .nrbc import NRBC
 from ..io.basicio import parallel_print as pprint
 from .lay_len import calc_size_lay
 from ..plots.plots_habc import plot_function_layer_size
-from ..tools.abc_labeling_cases import formatting_abc_layer_type, path_to_save_abc_case
+from ..tools.abc_set_path_cases import formatting_abc_layer_type, path_to_save_abc_case
 from ..utils.error_management import validate_enum, validate_numeric, validate_parameter
 from ..utils.freq_tools import freq_response
 from ..utils.typing import (AbsorbingBCsType, BoundaryConditionsType, HyperLayerDegreeType,
