+++ spyro/tools/abc_set_path_cases.py
@@ -1,6 +1,7 @@
-"""Methods to label the case study of the ABC scheme."""
+"""Methods to set paths for the case study of the ABC scheme (HABC, PML or NRBC)."""
 from numpy import pi
-from os import getcwd, makedirs, path
+from os import getcwd
+from pathlib import Path
 from ..io.basicio import parallel_print as pprint
 from ..utils.error_management import (mutually_exclusive_parameter_error, validate_enum,
                                       validate_numeric, validate_parameter, validate_string)
@@ -273,19 +274,19 @@ def path_to_save_abc_case(abc_type, abc_boundary_layer_shape=None, abc_deg_layer
         return path_save, case_nrbc, path_case_nrbc
 
 
-def create_folder(folder):
-    """Verify if a folder exists, if not, it creates the folder.
+def create_folder(folder: str | Path) -> None:
+    """Ensure a directory exists, creating it if needed.
 
     Parameters
     ----------
-    folder: `str`
-        Path to the folder to be created
+    folder : str or pathlib.Path
+        Path to the directory to create.
 
     Returns
     -------
     None
     """
-
-    # Create the folder if it does not exist
-    if not path.isdir(folder):
-        makedirs(folder)
+    # TODO: replace with the validation method to come
+    folder_path = Path(folder)
+    if not folder_path.is_dir():
+        folder_path.mkdir(parents=True, exist_ok=True)
